requirements.txt

feedparser
requests
PyYAML
jellyfish
beautifulsoup4
lxml
icalendar
python-dateutil
